Made small changes from gcc version.
I was using current incorrectly. I had a local array declated incorrectly. This could have been what was causing the errors in the output.

It was not, I have finished the error checking in draw tris so that is checks for very small triangles and not just triangles that are all 0s.

I didn’t initialize vertex current, did so.

reworking drawtris. Shifting singleton triangle logic to drawtris, made toFile() to actually draw a triangle

adding empty drawSquare()

added case 7